pyproject.toml
[tool.poetry]
name = "pptx"
version = "0.1.0"
description = ""
authors = ["infinage <37257700+Infinage@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-pptx = "^1.0.2"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"
jedi-language-server = "^0.41.4"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
